requirements.txt
langchain-groq
langchain
Pillow
chromadb
diffusers
sentence_transformers
accelerate